Merged

+++ iOS/G3MiOSSDK/Commons/Basic/CompositeElevationData.hpp
@@ -40,7 +40,7 @@ public:
 
   virtual ~CompositeElevationData() {
     size_t s = _data.size();
-    for (int i = 0; i < s; i++) {
+    for (size_t i = 0; i < s; i++) {
       if (_data[i] != NULL){
         _data[i]->_release();
       }
